Attempt to simplify class organization
