+++ auth/src/main/java/com/firebase/ui/auth/util/CredentialApiHelper.java
@@ -1,0 +1,11 @@
+package com.firebase.ui.auth.util;
+
+import com.google.android.gms.auth.api.credentials.Credential;
+import com.google.android.gms.common.api.Status;
+import com.google.android.gms.tasks.Task;
+
+public interface CredentialApiHelper {
+    Task<Status> disableAutoSignIn();
+
+    Task<Status> delete(Credential credential);
+}
